+++ plugins/apel/src/auditor_apel_plugin/publish.py
@@ -30,6 +30,8 @@ from auditor_apel_plugin.core import (
 )
 from auditor_apel_plugin.config import get_loaders, Config, MessageType
 
+TRACE = logging.DEBUG - 5
+
 
 def run(logger: Logger, config: Config, client):
     report_interval = config.plugin.report_interval
@@ -60,8 +62,6 @@ def run(logger: Logger, config: Config, client):
             begin_month = get_begin_current_month(current_time)
 
         for site in sites_to_report.keys():
-            logger.info(f"Getting records for {site}")
-
             if message_type == MessageType.individual_jobs:
                 start_time = get_start_time(config, time_dict, site)
                 logger.info(f"Getting records since {start_time}")
@@ -80,13 +80,13 @@ def run(logger: Logger, config: Config, client):
             filled_db = fill_db(config, db, message_type, field_dict, site, records)
             grouped_db = group_db(filled_db, message_type, optional_fields)
             message = create_message(message_type, grouped_db)
-            logger.debug(message)
+            logger.log(TRACE, message)
             signed_message = sign_msg(config, message)
-            # logger.debug(signed_message)
+            logger.log(TRACE, signed_message)
             encoded_message = base64.b64encode(signed_message).decode("utf-8")
-            # logger.debug(encoded_message)
+            logger.log(TRACE, encoded_message)
             payload_message = build_payload(encoded_message)
-            logger.debug(payload_message)
+            logger.log(TRACE, payload_message)
             post_message = send_payload(config, token, payload_message)
             logger.debug(post_message.status_code)
 
@@ -134,8 +134,8 @@ def main():
         "[%(asctime)s] %(levelname)-8s %(message)s (%(pathname)s at line %(lineno)d)"
     )
 
+    logging.addLevelName(TRACE, "TRACE")
     logging.basicConfig(
-        # filename="apel_plugin.log",
         level=log_level,
         format=log_format,
         datefmt="%Y-%m-%d %H:%M:%S",